// app/schema.ts
export type Database = {
// ...
  public: {
    Tables: {
      people: {
        Row: {
          created_at: string
          email: string | null
          firstname: string | null
          id: number
          surname: string | null
          user_id: string | null
        }
        Insert: {
          created_at?: string
          email?: string | null
          firstname?: string | null
          id?: number
          surname?: string | null
          user_id?: string | null
        }
        Update: {
          created_at?: string
          email?: string | null
          firstname?: string | null
          id?: number
          surname?: string | null
          user_id?: string | null
        }
        Relationships: []
      }
      questions: {
        Row: {
          created_at: string
          id: number
          question: string | null
          user_id: string | null
        }
        Insert: {
          created_at?: string
          id?: number
          question?: string | null
          user_id?: string | null
        }
        Update: {
          created_at?: string
          id?: number
          question?: string | null
          user_id?: string | null
        }
        Relationships: []
      }
// ...
      survey_people: {
        Row: {
          created_at: string
          id: number
          people_id: number | null
          survey_id: number | null
          user_id: string | null
        }
        Insert: {
          created_at?: string
          id?: number
          people_id?: number | null
          survey_id?: number | null
          user_id?: string | null
        }
        Update: {
          created_at?: string
          id?: number
          people_id?: number | null
          survey_id?: number | null
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "survey_people_people_id_fkey"
            columns: ["people_id"]
            isOneToOne: false
            referencedRelation: "people"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "survey_people_survey_id_fkey"
            columns: ["survey_id"]
            isOneToOne: false
            referencedRelation: "surveys"
            referencedColumns: ["id"]
          },
        ]
      }
      survey_questions: {
        Row: {
          created_at: string
          id: number
          question_id: number | null
          survey_id: number | null
          user_id: string | null
        }
        Insert: {
          created_at?: string
          id?: number
          question_id?: number | null
          survey_id?: number | null
          user_id?: string | null
        }
        Update: {
          created_at?: string
          id?: number
          question_id?: number | null
          survey_id?: number | null
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "survey-questions_question_id_fkey"
            columns: ["question_id"]
            isOneToOne: false
            referencedRelation: "questions"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "survey-questions_servey_id_fkey"
            columns: ["survey_id"]
            isOneToOne: false
            referencedRelation: "surveys"
            referencedColumns: ["id"]
          },
        ]
      }
      surveys: {
        Row: {
          appraisee_id: number | null
          created_at: string
          id: number
          name: string | null
          user_id: string | null
        }
        Insert: {
          appraisee_id?: number | null
          created_at?: string
          id?: number
          name?: string | null
          user_id?: string | null
        }
        Update: {
          appraisee_id?: number | null
          created_at?: string
          id?: number
          name?: string | null
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "survey_appraisee_fkey"
            columns: ["appraisee_id"]
            isOneToOne: false
            referencedRelation: "people"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}